refactor file-structure to separate MVC concerns
