+++ api/controllers/error-handling.controllers.js
@@ -14,10 +14,9 @@ exports.handlePSQLErrors = (err, req, res, next) => {
   if (err.code === "22P02") {
     res.status(400).send({ msg: "Invalid ID, must be a number" });
   }
-  if (err.code === '23502') {
+  if (err.code === "23502") {
     res.status(400).send({ msg: "Invalid format" });
-  }
-  else {
+  } else {
     next(err);
   }
 };
