+++ api/models/categories.models.js
@@ -1,0 +1,6 @@
+const db = require("../../db/connection");
+
+exports.fetchAllCategories = () => {
+  return db.query(`SELECT * FROM categories;`)
+  .then((res) => res.rows);
+};
