--- playground.sql
@@ -1,7 +1,0 @@
-\c nc_games_test
-
--- INSERT INTO comments(author, body, review_id)
--- VALUES ('Peter', 'Hello', 1)
--- RETURNING *;
-
-SELECT * FROM users;
